Update the resultsdb-ci-listener image to fix result publishing
No Fedora CI test results are being published to resultsdb at
the moment, because the 'version' property of the messages got
bumped to 1.1.14 and the listener only handled messages with
version strings starting 0.2. This bumps the listener image
to a new build which includes a patch that should make it handle
the 1.1.14-versioned messages (and is built on Fedora 39, not
37...) This is urgent because if the results aren't in resultsdb
Bodhi and greenwave don't know about them, and if they don't
know about them, updates get gated (dozens are stuck on this at
present).

Signed-off-by: Adam Williamson <awilliam@redhat.com>