haproxy: set keepalive mode for kojipkgs
We are hitting a sporadic and anoying 502 error with ostree pulls.
see https://pagure.io/releng/issue/11439

The problem seems to be between haproxy and varnish on kojipkgs01.

We set the httpclose option in haproxy globally, which closes
connections as soon as it thinks they are done.
Setting this option 'httpkeepalive' will keep connections alive
and handle the case of lots of fast connections downloading small
objects much better.

Sadly, we don't have a way to test this in staging, so we would need to
test in prod and roll back if there's problems.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>