batcave: renew proxy letsencrypt certs weekly
We occasionally run into problems with certs that aren't renewed in time
or are close to expiring. Just running the proxies playbook will renew
them, but in freezes or the like sometimes there's a long time period
where we don't run that playbook.

So, lets just run weekly with the right tag. This should renew any cert
thats close to expiring.

The job shouldn't normally output anything, but if there's errors it
will email them to admin@

Signed-off-by: Kevin Fenzi <kevin@scrye.com>