Revert "Reapply "Don't send a fedora-messaging message from the Pagure git hook""
This reverts commit 26358b62a2180520b5bd6f3a08710f9e2288eae6.

So, this broke the toddler that handles package retirements. ;(

It seems it was commited on feb 20th, but not actually pushed out until
March 4th. At that point, retirements stop. ;(

I think it's getting the new git pagure messages, but they don't
contain the information it expects so it just drops them.

We will need to adjust toddlers before we can re-drop this.