greenwave: update gating policy for openQA UEFI switchover
We're landing a big change to make openQA tests UEFI-by-default,
with BIOS as an occasional variant (exact opposite to the current
setup). This adjusts the gating policies appropriately. We have
to temporarily disable gating for a few tests, due to the fact
that there will be 'in-flight' updates which do not have a result
for the 'bios' scenario (they have results for 'uefi' and '64bit',
with '64bit' being a BIOS result; post-switchover, updates will
have a result for '64bit' and 'bios', with '64bit' being a UEFI
result). I will re-enable this in about a month and clean up any
stragglers somehow.

Signed-off-by: Adam Williamson <awilliam@redhat.com>