greenwave: don't gate c-p-standard on container tests
This was wrong and I'm surprised we never noticed till now. We
don't schedule these tests on critical-path-standard so we should
not gate on them. The alternative fix is to keep gating and start
testing, but the only thing that's actually in c-p-standard is
fprintd-pam and we don't pull fprintd stuff into containers, so
there's no sense testing it, really.

Signed-off-by: Adam Williamson <awilliam@redhat.com>