fix(os:cloud-image-uploader): set workingDir on container
The latest import failed due to "ERROR: Unable to create local
directories(/.ansible/tmp): [Errno 13] Permission denied: b'/.ansible'".
Which implies the code is being executed from `/`, despite the WORKDIR
variable being set in the container image--I suspect this is a quirk of
kube/openshift that was not expected.

This change sets the workingDir to /srv/cloud-uploader, as specified in
the Containerfile, which should resolve the execution error.

Signed-off-by: Neil Hanlon <neil@shrug.pw>