bodhi: set primary arches for current stable releases
bodhi is failing to finish a f40-updates-testing push because it's
looking for the ppc64le repomd on the primary location instead of the
secondary one.

This should fix it, but I don't understand how it worked before.
Did we just get lucky and always check aarch64 before anything else?
I can't see anything that changed in config around this off hand. ;(

Will try this and see if it fixes it for now...

Signed-off-by: Kevin Fenzi <kevin@scrye.com>