Bodhi compose config: build ostrees and containers in variants
Our compose attempt with the containers being built under
Everything didn't go as planned because the filenames of the
container images turn out to be determined by the variant, so
they conflicted and only the last one built survived. So, we
really need to build them under unique variants for now. A
previous commit added the variant definitions.

We don't *need* to build the ostrees under the variants, but it
feels more consistent this way, both with the containers and
with how it works in the Rawhide/Branched composes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>