bodhi / createerepo / epel8: try and get primary back on xz
We changed around compression as part of a Fedora change and move to
createrepo_c > 1.0, but the epel8 repos are causing a bit of problem for
users who use mergerepo_c on them from older (<1.0) createrepo_c.
They can't understand the primary.xml being zstd compressed.
So, lets try and set this to switch primary back to xz.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>