greenwave: stop gating f39 updates for CoreOS
Many of the packages currently gated by the CoreOS tests are also used
outside of the CoreOS artifacts. And so it happens that e.g. Bodhi
updates are opened for releases that CoreOS itself no longer uses.

However, since those releases aren't monitored by CoreOS anymore, no
CoreOS tests run. In that case, we shouldn't be gating updates on those
releases.

Fedora 39 is in this scenario right now. Ideally, we could use symbolic
names here instead like `fedora-latest-stable` or something. For now,
I'll add this to a rebase checklist we have to make sure we update this
list shortly after a GA release.