openqa/dispatcher: add a cron job to send missed test results
This works around an annoying problem where, for some reason, we
sometimes just miss sending completed test results to resultsdb.
I've never been able to figure out why this happens, but this
should band-aid it by looking, daily, for updates stuck in
waiting gating status, checking for cases where a test finished
but we didn't send a result, and sending it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>