buildhw-a64: move 4 old mt snow servers to buildhw
We have 4 of the new mt snow boxes that were bvmhosts before, but we
moved the vm's to the newer generation versions, so we should use these
as buildhw boxes. I plan to add 2 of them to the runroot channel for
composes and 2 of them general builders in the heavybuild channel to
help with chromium builds.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>