Update owner-sync-pagure.j2 to work with EPEL 10, second attempt
Originally attempted in 01be34a70645eb744751b5341fd3a2339d70433c, but it
didn't work correctly and needed to be reverted.  Some notable
differences this time around:

- Previously the script predicted branch and version based on the tag.
I had this switched to get both of those from the bodhi release, but not
all tags have a corresponding bodhi release.  So this now takes a hybrid
approach of using bodhi release if available, and falling back to the
prediction approach otherwise.
- Retired package lists from the lookaside are now skipped if they don't
exist, both during the retrieval and when checked later.
- The get_project_branches function now returns a set.

Signed-off-by: Carl George <carlwgeorge@gmail.com>