Fix RuntimeError in owner-sync-pagure.j2
This is a small follow up fix related to 01be34a70645eb744751b5341fd3a2339d70433c.
Previously the script modified a list while iterating through it, which
is always a bad idea, but it didn't throw an error.  My last changes
switched it to iterating through a list, which surfaced the flaw as a
RuntimeError.  Using a set comprehension is a cleaner way to generate
the set we need to return.

Signed-off-by: Carl George <carlwgeorge@gmail.com>