bodhi: Remove usage of version_int from pungi template
EPEL 10 has minor versions, and bodhi's Release.version_int property
doesn't handle versions with a dot correctly (yet).  Almost all of our
usage is for constructing strings or for equality comparisons that can
be strings, so we can just use Release.version instead.  There is an
instance of a numerical greater than or equal to comparison, but it is
inside a condition for release.id_prefix == 'FEDORA', so it shouldn't
get evaluated for EPEL updates.

Signed-off-by: Carl George <carlwgeorge@gmail.com>