bodhi: Fix nested quotation
In c7fbc26afe9f770414a49d150d9072c1e58b21ec I switched an integer 41 to
a string, but didn't notice that it was inside another string as part of
the double templating this file does.  This results in a template error:

jinja2.exceptions.TemplateSyntaxError: expected token 'end of statement block', got 'integer'

Other strings inside this bigger string use double instead of single
quotes to handle this, so let's take the same approach on this line.

Signed-off-by: Carl George <carlwgeorge@gmail.com>