[toddlers] Update distgit_bugzilla_sync config
I changed the logic a little to prevent the issue with container projects that
happened before, but this needs new change in configuration.

The distgit_bugzilla_sync will now map project to product first based on
namespace and if there are more Bugzilla products mapped to same namespace it
will use the branch_regex. This should be more clear and prevent the container
issue in future.

Signed-off-by: Michal Konecny <mkonecny@redhat.com>