fedora-image-uploader: Drop unused config options for AWS images
The AMI description setting wasn't actually being used.

More importantly, however, is the AMI volume size is now also unused.
The reason for this is that when we import the image, the default is to
use the snapshot's size as the volume size, but you can optionally set
it to something else. AWS pre-allocates volumes of a couple different
sizes (currently 1G, 5G, 8G, and 10G).

Folks building the image set the size to be 5G, but this setting
(carried over from fedimg) overrode it. Dropping it lets them control
the AMI size by adjusting how big the images are. Probably not optimal
for upload speed, but less confusing since there are fewer configuration
layers people might not be aware of.