[mailman] Use xapian as a backend for fulltext search
The whoosh engine doesn't seem to be maintained anymore and when generating the
index it got stuck when the size was about 20 GB. Xapian on the other hand looks
more stable and much quicker when generating the index.

Unfortunately the xapian-haystack package is not available in Fedora/EPEL yet
and waiting for review to be finished.

Link to review bug: https://bugzilla.redhat.com/show_bug.cgi?id=2313507

Signed-off-by: Michal Konecny <mkonecny@redhat.com>