Revert "fedora-image-uploader: Add staging config for container uploads"
This reverts commit 5e0ad1134d7ed7796f8385d6960edfc1ef16d31a (pr #2200).

Unfortunately, using buildah inside an unprivileged OpenShift container
turns out to not be very simple, even though we're not building any
containers, just importing and pushing them.

We can either figure out how to make it work with OpenShift (and while
it is definitely possible, I don't know if folks are okay with the
compromises that might be required) or deploy it in a VM for now.

In the meantime, the staging container is bootlooping so I'd like to
back this configuration out for the sake of my inbox.