[ipa/server] Change how the check for ipa-replica-install is being done
As found on in https://pagure.io/fedora-infrastructure/issue/12149 the current
method of checking if the ipa-replica-install have to be executed is not
reliable enough.

Let's instead check for log file that is created during the installation. This
will need manual intervention in case the ipa-replica-install needs to be run
again, but it will be much more reliable otherwise.

Signed-off-by: Michal Konecny <mkonecny@redhat.com>