koji_hub: do not send buildSRPMFromSCM to createrepo channel
Long ago we sent buildSRPMFromSCM tasks to createrepo channel because we
didn't want s390x (which wasn't in that channel) to do them. However,
with koji 1.35.x these tasks obey the noarch_arches tag value as to
which arch they run on. This means we would need to add s390x and
ppc64le to createrepo for them to always work.

So, instead lets drop this hub policy for buildSRPMFromSCM and then they
will just get default channel and obey noarch_arches as is expected.
If s390x proves too slow at these tasks, we can just remove it from
noarch_arches or come up with another plan.

newRepo distRepo in this policy was pointless, koji automatically does
these tasks with channel=createrepo, so no need to specify it here.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>