📧 roles(fasjson): Create jww@ email alias for @jflory7
This commit adds a new fedoraproject.org email alias for me, using my
new initials. The short story is that I am in the process of legally
changing my last name. Although the name change is still underway, I
want to prepare the infrastructure bits in advance so I can have a
graceful transition once the legal process is complete.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>