bodhi/backend: use major version to construct EPEL latest link
The latest epel-release symlink name is constructed by parsing the fifth
digit out of the release.  That was never going to work when we reached
a double digit major version.  It is resulting in a symlink of
epel-release-latest-1.noarch.rpm for epel10.0.  Instead, let's switch it
to use the version value (without the minor version) from the release.

Signed-off-by: Carl George <carlwgeorge@gmail.com>