osbuildapi-update: update the osbuildapi ipset atomically
Prior this commit, the script flushed the content of the osbuildapi
set before adding the newly resolved IP addresses into it. Even worse,
the DNS resolve for sso.redhat.com happened after the flush, so there
was no IP address for sso.redhat.com for quite some time (resolving
sso.redhat.com without caching takes 30 ms on my machine).

This commit fixes that by introducing a secondary ipset that is firstly
filled with the resolved IP addresses, and then its content is swapped
with the main set. This should hopefully reduce the likelihood of
random races.