Ensure python3-dnf is installed on Fedora 41+ hosts
dnf5 requires neither python3-rpm nor python3-dnf, so we cannot
assume they are installed (as they always were before). ansible's
package_facts needs python3-rpm (even in the latest upstream
version), so we at least need that. The ansible in rhel9 (which
is what's on batcave01 currently) does not support dnf5 at all,
so we need python3-dnf installed so it can use its dnf3/4
support. python3-dnf relies on python3-rpm, so we can just
check for and install python3-dnf. Once batcave01 is updated to
an ansible that supports dnf5 we won't need python3-dnf, but we
*will* still need python3-rpm.

Stuffing this in yumrepos.yml as it runs on everything, early,
and is packaging-related. We could make it a separate task but
then we'd need to make every playbook include it...

Signed-off-by: Adam Williamson <awilliam@redhat.com>