koji_builder: use iptables backend for libvirt
In f41+ libvirt defaults to using nftables if both it and iptables
are installed, but it doesn't seem to work with imagefactory/oz
virt instances and our iptables setup.
So, lets revert back to iptables for now.
We can switch back if we can fix the incompatiblity, switch builders
to nftables, or stop using oz/IF

Signed-off-by: Kevin Fenzi <kevin@scrye.com>