greenwave: templatize gating config
This uses the release number templates for the product_versions
blocks in the gating config, instead of hardcoding them. I tried
to include some flexibility so we can easily manually flip gating
on or off for Rawhide or Branched - for specific rules, or the
whole file - around branch time, if we need to (sometimes we do,
as we can get in catch-22 situations and stuff during branching).

Signed-off-by: Adam Williamson <awilliam@redhat.com>