copr: the eth0 device is now called ens5
At first, I wanted to fix the current name and change it back to eth0 but it
seems this is a desired naming from now on and we should probably adopt it.

https://fedoraproject.org/wiki/Changes/EnableConsistentDeviceNamingCloud

This commit will probably break playbooks for the production instances but I
cannot do a "if devel" condition here. So hopefully, we will be fine.