roles/bodhi2: Set EPEL to use zstd repodata and disable unusable features
Zstd repodata support was introduced in DNF quite some time ago, and
is quite a bit faster and less resource intensive.

Additionally, there is no point in producing sqlmd or DeltaRPMs when
support for both has been removed from RHEL for a fair bit of time.

Regardless, we should leave things mostly as-is for EPEL 9, and thus
we retain older xz compression settings for that.

Signed-off-by: Neal Gompa <ngompa@fedoraproject.org>