releng: drop clean-ami's job
This job was used to cleanup old ami's from fedimg.
It then reported to a pagure project about what it did.
However, it's token has been expired and it no longer needs to exist.
Garbage collection of ami's should take place in the new
cloud-image-uploader (well, it marks when things are eol).
So, lets just remove this old and nolonger needed cron.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>