buildvm_s390x: cleanup and add 7 new builders.
We no longer need to keep any config around from boston, so nuke that.
All the buildvm's are now kvm, so we don't need a seperate group/config
for that, so remove it.

Consolidate the variables down to buildvm_s390x group and only have ip
address for host vars.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>