batcave: redirect http to https
We hadn't done this in the past because virt-install couldn't do https
urls, but thats long since no longer the case.

So, move all the ks_url and ks_repo calls to use https and since they
are using https, they need to use the valid
infrastructure.fedoraproject.org name so they can verify the cert.

Set batcave to redirect http to https now since that was the thing
holding this back.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>