🔧 proxies-redirects: Redirect parameter for docs.fp.o/dei/ redirect
This commit enhances the redirect regex used for the docs.fp.o redirect
for the DEI Team documentation. I follow @kevin's example from the CPE
to CLE docs redirect, which uses a parameter and redirects a specific
page to the new address. Currently, the DEI Team docs redirect will take
the user to the main page of the team, regardless of what specific page
is visited.

I also added a `docs` tag to the CLE tags redirect, since it is a Fedora
Docs-specific redirect too.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>