bastion: fix delivering non contributors emails locally
Should fix:
https://pagure.io/fedora-infrastructure/issue/12361

Basically postfix default is:
local_recipient_maps = proxy:unix:passwd.byname $alias_maps
so if the user is a local user or an alias, it's valid.
However, sssd and ipa show all users (even ones with no
access to that host). This means we were accepting and delivering
(locally) emails for anyuser@fedoraproject.org.

Setting this to just $alias_maps will just treat aliases as valid
and ignore all the local users. This should be fine as we use
aliases to send even to root or other system users.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>