Tweak coreos block in greenwave config to avoid newline issues
Well, I hope, anyway. This is a nightmare to predict/test because
jinja is weird and ansible puts it in a weird non-standard mode.
But this does the same thing as the product_versions macro we
use for all other releases - just leave out all attempts to avoid
blank lines.

It *might* cause blank lines in some states, but I *think* I found
(while testing product_versions) that blank lines don't break the
parsing. I can't find anywhere in the YAML spec where it actually
defines when a block sequence *ends*, but as a matter of fact it
seems like whatever parser is being used here doesn't choke on
blank lines.

My ideal goal here is that, whatever entries are present or
absent based on the conditionals, we get a clean block sequence
with each entry on its own line and no blank lines, but this
seems to be *incredibly* hard to get jinja to do.

Signed-off-by: Adam Williamson <awilliam@redhat.com>