🔧 roles(fasjson): Add foa,jasonbrooks to sponsors alias
This commit adds the `foa@` alias (AKA @amoloney) and @jasonbrooks to
the email address used for managing Flock to Fedora sponsor comms. Aoife
is already involved with the Flock organizing and it would be helpful
for her to see emails sent to this alias. I am adding Jason to the alias
for transparency and to provide me coverage during periods when I am
away or out of office.

@mattdm, @dvolavko, and @shaunm retain their access to emails sent to
the alias.

Signed-off-by: Justin W. Wheeler <jwf@redhat.com>