Belatedly lift F41 Final freeze, remove unused RelEngFrozen
It seems this has been set true ever since F41 Final freeze. It
probably should have been set false after F41 shipped, but we
missed it.

We did freeze for F42 branching, but I'm pretty sure we declared
that over now. Next freeze is on Feb 18, for Beta.

Also, the RelEngFrozen variable is no longer used by anything.
It was only ever used for one thing in Bodhi config, but that
use was removed in 02cdf36 . So let's get rid of the variable.

Signed-off-by: Adam Williamson <awilliam@redhat.com>