greenwave config changes related to kiwi
The kiwi_build test for containers has been around long enough
that we can require it now. Also, drop live_build requirement
for Workstation because we're going to flip openQA to using Kiwi
to build the Workstation live (just as releng has for the
official one). This doesn't risk allowing broken updates through
as we still require the install_default_update_live tests to
pass, and they can't pass if the image build fails.

Signed-off-by: Adam Williamson <awilliam@redhat.com>