osbuildapi-update: just keep adding to ipset
It seems sso.redhat.com has a 20 second time to live, so it changes
often. However, it seems to only change the pool of ips from a small
(5-10) number. So, lets just add to the set so we allow any of the ips
we have seen. The set will grow, but if the pool is small it should
stablize to have all the ips in the pool.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>