Clarify the Frozen variable
Per https://pagure.io/infra-docs-fpo/pull-request/361 there seems
to be substantial confusion about what the "Frozen" variable
means: does it refer to the infrastructure freeze, or a freeze
for the pending release? It seems like @kevin thought about it
as referring to the former, but the only practical use of it in
this repo treats it as the latter.

Let's make this clear with explicitly-named variables.
InfraFrozen tracks the infrastructure freeze, NextReleaseFrozen
tracks freezes for the pending release. We formerly had a var
called RelEngFrozen which had a similar meaning, but I like this
name more because the freeze doesn't apply to *all* releng
operations - just because the current Branched is frozen doesn't
mean Rawhide or stable releases are frozen.

We keep Frozen for 'backwards compatibility', in case anyone is
referring to this variable from outside of the repo.

This goes along with https://pagure.io/infra-docs-fpo/pull-request/361
which updates the SOPs.

Signed-off-by: Adam Williamson <awilliam@redhat.com>