[pagure] Tweak the performance of httpd server
Today the pagure started to be unresponsive. After few hours of
debugging we narrowed the issue to worker limit. We tweaked the value to
something that should be more responsive and it helped.

Here is the PR to make the change permanent.

Thanks @gwmngilfen and @nphillip for help with that.