batcave / ansible-server: specifcy any local ansible collections
We want to specify the exact collections we install and the version so
that we know whats installed and what version it is.
Ideally we would install all the collections via rpms.
When thats not possible, we should install them via ansible
so they are tracked and known, and we can update them as needed.

Any other collections locally installed that I couldn't find a caller
for have been removed. We can easily add them here if there's more we
want to install.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>