bvmhost-p09-04/05: no nbde here
I had reinstalled these both with no encryption in an attempt to see if
I could get more performance from them. Since we moved to iscsi this is
moot, and we should probibly reinstall them like the others again, but
for now just disable nbde so everything works with the playbook.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>