fedora-image-uploader: fix deployment trigger on image changes
I broke the trigger when I switched from one container to multiple
containers in the pod. Syntax-wise, I found this multi-line variant in
the CoreOS Cincinnati deployment config, and it seems like there's not a
way to say "all container images in the spec". Or there might be, but I
couldn't find an example or documentation.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>