zabbix / templates: just use inventory_hostname not short
This causes an issue where we have a proxy01.iad2 and a proxy01.rdu3.
They are different hosts, but if it uses the short name then the second
one fails because it has the same name as the first.

Lets just use fqdn for now and see if causes any problems in
display in the interface. If it does, we can change it to
perhaps include the second octet.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>