copr-be-dev: remove old, commented-out builder images
Not worth tracking n-0 and n-1 images for the STG instance. In the rare
situation when rolling-back is needed, we have a git history. Also, we cannot
remove the older images from AWS just yet, because they are still referenced by
n-1 comment in the production config. That's why I am leaving the
warning only in production.