proxies / rdu3: set rdu3 proxies to proxy to the rdu3 openshift cluster
The iad2 and rdu3 openshift clusters have the same name
(yes, I know, I was trying to be clever!)
So, in rdu3 we want *.apps.ocp.fedoraproject.org to proxy to the rdu3
cluster and everywhere else to the iad2 cluster (until we move).
This should point proxy01.rdu3 and proxy10.rdu3 the right way.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>