📧 roles(fasjson): Update `codeofconduct` alias for current membership
This commit changes the `codeofconduct` email alias to account for the
departure of Matthew Miller (@mattdm) and Marie Nordin (@riecatnor) as
members of the Code of Conduct Committee. Additionally, @jspaleta will
eventually receive mail sent to this alias after the `fpl` alias is
updated. Also, Dorka Volavkova (@dvolavko) is retiring as a full member
of the Fedora Code of Conduct Committee, and instead, will serve as a
member of the Flock to Fedora on-site response team, along with several
other appointees for the purpose of Flock.

Additionally, @jberkus and @nimbinatus are added as temporary members
until September 30th, 2025 on the Fedora Code of Conduct Committee, to
fill in for the recent departures and my planned leave of absence later
this month.

ref: https://docs.fedoraproject.org/en-US/project/code-of-conduct/committee/
ref: https://pagure.io/CoC/issue/355

Signed-off-by: Justin Wheeler <jwheel@redhat.com>