buildvm-ppc64le.rdu3: deploy ppc64le builders in rdu3
Note that these don't do nested virt due to the way we have the host
setup. We are working on changing that, but for now, this should work
fine for normal builds. Only parts of composes that need nested virt
will fail for a week or so until we get a power9 moved to the new dc
or reconfigure a power10 box.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>