pkgs: create a local pagure user.
Seems like we set this up many years ago for staging, but then we added
a comment to do it for prod and never did.

This is needed in todays rdu3 setup. The pagure user needs to be local
so it can sudo with the local sudo config instead of the ipa sudo
config.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>