inventory: add ipv6 addresses to various hosts that need them
We don't have ipv6 routing setup yet, but are scheduled to work on that
soon. To get ready for that, lets add ipv6 addresses to the (few)
machines that will actually need them.

We do not want to add ipv6 to all hosts. The vast majority of them never
need to talk to the outside world directly and shouldn't have a ipv6
address that can do this.

These few hosts are ones with external nat mappings where it is
desireable that they be able to handle ipv6 connections.

Note that we also do NOT want to add any of these to dns until
they are known working. We also will likely have to adjust nftables
to allow the services on ipv6 that we do on ipv4 (if they make sense).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>