Drop broken ns03 from openQA worker network config temporarily
This server isn't working and we can't figure out why not. It's
a problem for openQA because we copy the host's DNS config into
'advanced networking' openQA guests, and then when we do a
FreeIPA deployment test, it picks up both DNS servers, tries to
confirm both work, and fails. So we need to take ns03 out until
it's fixed.

Signed-off-by: Adam Williamson <awilliam@redhat.com>