iscsi_client: readd role, apply to power10 host and switch guests to use it
This re-adds a iscsi_client role we had in iad2 back in in rdu3.
When then apply it to bvmhost-p10-01 to login and use a iscsi lun from
the rdu3 netapp. We then move the buildvm-ppc64le vm's to use this iscsi
volume instead of local storage.

As we reinstall those builders they will use the iscsi volume.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>